+++ lib/avs_image.dart
@@ -13,7 +13,7 @@ part 'functions/_check_svg.dart';
 
 enum AVSImageType { network, asset, file }
 
-enum CloseButtonPosition { topRight, topLeft, bottomRight, bottomLeft }
+enum ButtonPosition { topRight, topLeft, bottomRight, bottomLeft }
 
 enum ZoomStyle { onTap, onDoubleTap, onLongPress }
 
@@ -140,19 +140,11 @@ class AVSImage extends StatelessWidget {
           child: AspectRatio(
             aspectRatio: 1,
             child: ClipRRect(
-              borderRadius: isCircle == true
-                  ? BorderRadius.circular(360)
-                  : radius ?? BorderRadius.zero,
+              borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
               child: GestureDetector(
-                onTap: (onTap == null && zoom == false)
-                    ? null
-                    : () async => _onTapFunction(context),
-                onLongPress: (onLongPress == null && zoom == false)
-                    ? null
-                    : () async => _onLongPressFunction(context),
-                onDoubleTap: (onDoubleTap == null && zoom == false)
-                    ? null
-                    : () async => onDoubleTapFunction(context),
+                onTap: (onTap == null && zoom == false) ? null : () async => _onTapFunction(context),
+                onLongPress: (onLongPress == null && zoom == false) ? null : () async => _onLongPressFunction(context),
+                onDoubleTap: (onDoubleTap == null && zoom == false) ? null : () async => onDoubleTapFunction(context),
                 child: _buildBody(),
               ),
             ),
@@ -160,38 +152,22 @@ class AVSImage extends StatelessWidget {
         );
       } else {
         return ClipRRect(
-          borderRadius: isCircle == true
-              ? BorderRadius.circular(360)
-              : radius ?? BorderRadius.zero,
+          borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
           child: GestureDetector(
-            onTap: (onTap == null && zoom == false)
-                ? null
-                : () async => _onTapFunction(context),
-            onLongPress: (onLongPress == null && zoom == false)
-                ? null
-                : () async => _onLongPressFunction(context),
-            onDoubleTap: (onDoubleTap == null && zoom == false)
-                ? null
-                : () async => onDoubleTapFunction(context),
+            onTap: (onTap == null && zoom == false) ? null : () async => _onTapFunction(context),
+            onLongPress: (onLongPress == null && zoom == false) ? null : () async => _onLongPressFunction(context),
+            onDoubleTap: (onDoubleTap == null && zoom == false) ? null : () async => onDoubleTapFunction(context),
             child: _buildBody(),
           ),
         );
       }
     } else {
       return ClipRRect(
-        borderRadius: isCircle == true
-            ? BorderRadius.circular(360)
-            : radius ?? BorderRadius.zero,
+        borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
         child: GestureDetector(
-          onTap: (onTap == null && zoom == false)
-              ? null
-              : () async => _onTapFunction(context),
-          onLongPress: (onLongPress == null && zoom == false)
-              ? null
-              : () async => _onLongPressFunction(context),
-          onDoubleTap: (onDoubleTap == null && zoom == false)
-              ? null
-              : () async => onDoubleTapFunction(context),
+          onTap: (onTap == null && zoom == false) ? null : () async => _onTapFunction(context),
+          onLongPress: (onLongPress == null && zoom == false) ? null : () async => _onLongPressFunction(context),
+          onDoubleTap: (onDoubleTap == null && zoom == false) ? null : () async => onDoubleTapFunction(context),
           child: _buildErrorWidget(context),
         ),
       );
@@ -215,23 +191,18 @@ class AVSImage extends StatelessWidget {
   }
 
   Widget _buildAssetSVG() {
-    return gradient == null
-        ? _buildLocalSVGNoGradient()
-        : _buildLocalSVGWithGradient();
+    return gradient == null ? _buildLocalSVGNoGradient() : _buildLocalSVGWithGradient();
   }
 
   Widget _buildLocalSVGNoGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: SvgPicture.asset(
         url,
         height: height,
         width: width,
         fit: _defaultFit,
-        colorFilter:
-            color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
+        colorFilter: color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
         alignment: alignment,
       ),
     );
@@ -239,12 +210,9 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildLocalSVGWithGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: _SVGLinearGradientMask(
-        gradient: gradient ??
-            const LinearGradient(colors: [Colors.black, Colors.black26]),
+        gradient: gradient ?? const LinearGradient(colors: [Colors.black, Colors.black26]),
         child: SvgPicture.asset(
           url,
           height: height,
@@ -258,23 +226,18 @@ class AVSImage extends StatelessWidget {
   }
 
   Widget _buildNetworkSVG() {
-    return gradient == null
-        ? _buildNetworkSVGNoGradient()
-        : _buildNetworkSVGWithGradient();
+    return gradient == null ? _buildNetworkSVGNoGradient() : _buildNetworkSVGWithGradient();
   }
 
   Widget _buildNetworkSVGNoGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: SvgPicture.network(
         url,
         height: height,
         width: width,
         fit: _defaultFit,
-        colorFilter:
-            color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
+        colorFilter: color == null ? null : ColorFilter.mode(color!, BlendMode.srcIn),
         alignment: alignment,
       ),
     );
@@ -282,12 +245,9 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildNetworkSVGWithGradient() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: _SVGLinearGradientMask(
-        gradient: gradient ??
-            const LinearGradient(colors: [Colors.black, Colors.black26]),
+        gradient: gradient ?? const LinearGradient(colors: [Colors.black, Colors.black26]),
         child: SvgPicture.network(
           url,
           height: height,
@@ -302,9 +262,7 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildAssetImage() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: Image.asset(
         url,
         color: color,
@@ -321,16 +279,12 @@ class AVSImage extends StatelessWidget {
   }
 
   Widget _buildNetworkImage() {
-    return cachedImage == true
-        ? _buildNetworkImageWithCache()
-        : _buildNetworkImageNoCache();
+    return cachedImage == true ? _buildNetworkImageWithCache() : _buildNetworkImageNoCache();
   }
 
   Widget _buildNetworkImageNoCache() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: Image.network(
         url,
         color: color,
@@ -344,10 +298,7 @@ class AVSImage extends StatelessWidget {
           } else {
             return showProgressIndicator
                 ? _buildProgressIndicator(
-                    ((loadingProgress.cumulativeBytesLoaded) /
-                                (loadingProgress.expectedTotalBytes ?? 1))
-                            .toDouble() *
-                        100.toDouble(),
+                    ((loadingProgress.cumulativeBytesLoaded) / (loadingProgress.expectedTotalBytes ?? 1)).toDouble() * 100.toDouble(),
                   )
                 : const SizedBox();
           }
@@ -362,9 +313,7 @@ class AVSImage extends StatelessWidget {
 
   Widget _buildNetworkImageWithCache() {
     return ClipRRect(
-      borderRadius: isCircle == true
-          ? BorderRadius.circular(360)
-          : radius ?? BorderRadius.zero,
+      borderRadius: isCircle == true ? BorderRadius.circular(360) : radius ?? BorderRadius.zero,
       child: CachedNetworkImage(
         imageUrl: url,
         height: height,
@@ -373,10 +322,7 @@ class AVSImage extends StatelessWidget {
         fit: _defaultFit,
         maxHeightDiskCache: 3000,
         color: color,
-        progressIndicatorBuilder: (context, url, progress) =>
-            showProgressIndicator
-                ? _buildProgressIndicator(progress.downloaded.toDouble())
-                : const SizedBox(),
+        progressIndicatorBuilder: (context, url, progress) => showProgressIndicator ? _buildProgressIndicator(progress.downloaded.toDouble()) : const SizedBox(),
         errorWidget: (context, url, error) {
           _showLog("Network Image With Cache Exploded: $url");
 
@@ -463,16 +409,10 @@ ImageProvider AVSImageProvider(
   LinearGradient? gradient,
 }) {
   bool isSvg = _isSvgCheck(path);
-  bool isLocalPosition =
-      _getImageType(path) == AVSImageType.network ? false : true;
+  bool isLocalPosition = _getImageType(path) == AVSImageType.network ? false : true;
 
   if (isSvg) {
-    return AVSSVGProvider(path,
-        color: color,
-        scale: scale,
-        height: height,
-        width: width,
-        gradient: gradient);
+    return AVSSVGProvider(path, color: color, scale: scale, height: height, width: width, gradient: gradient);
   } else if (isSvg == false && isLocalPosition) {
     return AssetImage(path);
   } else if (isSvg == false && !isLocalPosition) {